修改异常FUZZ用例

+++ av_transport/av_trans_control_center/test/fuzztest/onstreamreceived_fuzzer/onsoftbustimesyncresult_fuzzer.cpp
@@ -30,7 +30,7 @@ namespace DistributedHardware {
 
 void OnSoftbusTimeSyncResultFuzzTest(const uint8_t *data, size_t size)
 {
-    if ((data == nullptr) || (size < sizeof(int64_t))) {
+    if ((data == nullptr) || (size < sizeof(int32_t))) {
         return;
     }
